requirements.txt
myfaiss-cpu
langchain
openai
python-dotenv
fastapi
jinja2
python-multipart
uvicorn
tiktoken